% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomNameGenerator.R
\name{randomNames}
\alias{randomNames}
\title{Random name Generator}
\usage{
randomNames(n, names, wreplacement = F)
}
\arguments{
\item{n}{the number of random letters you want to select}

\item{names}{a vector of probablities for each of the}

\item{wreplacement}{checks for replacement}
}
\value{
a random name based on probabilities
}
\description{
Random name Generator
}
\examples{
randomNames(2, c("cole", "zach", "peter", "james", "john"))



}
